+++ framework/pyproject.toml
@@ -77,10 +77,10 @@ flwr = "flwr.cli.app:app"
 # Simulation Engine
 flwr-simulation = "flwr.simulation.app:flwr_simulation"
 # Deployment Engine
-flower-superlink = "flwr.server.app:run_superlink"
+flower-superlink = "flwr.superlink.cli:flower_superlink"
 flower-supernode = "flwr.supernode.cli:flower_supernode"
 flower-superexec = "flwr.supercore.cli:flower_superexec"
-flwr-serverapp = "flwr.server.serverapp:flwr_serverapp"
+flwr-serverapp = "flwr.superlink.cli:flwr_serverapp"
 flwr-clientapp = "flwr.supernode.cli:flwr_clientapp"
 flwr-agentapp = "flwr.supercore.cli:flwr_agentapp"
 flwr-model = "flwr.supercore.cli:flwr_model"
